re-organize state
